Call inflate with Z_FINISH and handle Z_STREAM_END
According to zlib doc: "inflate() should normally be called until it returns
Z_STREAM_END or an error. However if all decompression is to be performed in a
single step (a single call of inflate), the parameter flush should be set to
Z_FINISH. In this case all pending input is processed and all pending output is
flushed;".